+++ mycellar-java/mycellar-domain/src/main/java/fr/peralta/mycellar/domain/contact/repository/ContactRepository.java
@@ -48,7 +48,7 @@ public interface ContactRepository extends
      * @param count
      * @return
      */
-    List<Contact> getLastContacts(ContactOrder orders, int first, int count);
+    List<Contact> getLastContacts(ContactOrder orders, long first, long count);
 
     /**
      * @param producer
@@ -64,7 +64,7 @@ public interface ContactRepository extends
      * @param count
      * @return
      */
-    List<Contact> getAllForProducer(Producer producer, ContactOrder orders, int first, int count);
+    List<Contact> getAllForProducer(Producer producer, ContactOrder orders, long first, long count);
 
     /**
      * @param producer
