+++ mycellar-java/mycellar-domain/src/main/java/fr/peralta/mycellar/domain/shared/repository/EntityRepository.java
@@ -39,6 +39,6 @@ public interface EntityRepository<E extends IdentifiedEntity, OE, O extends Abst
      * @param count
      * @return
      */
-    List<E> getAll(O orders, int first, int count);
+    List<E> getAll(O orders, long first, long count);
 
 }
