+++ mycellar-java/mycellar-interfaces/mycellar-interface-facade-client/src/main/java/fr/peralta/mycellar/interfaces/facades/stack/StackServiceFacade.java
@@ -49,7 +49,7 @@ public interface StackServiceFacade {
      * @param count
      * @return
      */
-    List<Stack> getStacks(StackOrder orders, int first, int count);
+    List<Stack> getStacks(StackOrder orders, long first, long count);
 
     /**
      * @param stackId
